+++ services/dataplatform/versions.go
@@ -13,7 +13,13 @@ import (
 func (c *Client) GetVersions(ctx context.Context) ([]string, *dataplatform.APIResponse, error) {
 	versions, apiResponse, err := c.sdkClient.DataPlatformMetaDataApi.VersionsGet(ctx).Execute()
 	apiResponse.LogInfo()
-	return versions, apiResponse, err
+	if err != nil {
+		return nil, apiResponse, err
+	}
+	if versions.Items == nil {
+		return nil, nil, fmt.Errorf("expected a list of Dataplatform versions but received 'nil' instead")
+	}
+	return *versions.Items, apiResponse, err
 }
 
 func SetVersionsData(d *schema.ResourceData, versions []string) diag.Diagnostics {